improvements in Freemarker Config
