+++ src/main/java/io/github/gasparbarancelli/blog/ui/IndexController.java
@@ -1,0 +1,16 @@
+package io.github.gasparbarancelli.blog.ui;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/")
+public class IndexController {
+
+    @GetMapping
+    public String index() {
+        return "index";
+    }
+
+}
