+++ src/main/resources/static/css/index.css
@@ -1,0 +1,4 @@
+h1 {
+    color: green;
+    font-family: 'Raleway', sans-serif;
+}
