+++ src/main/resources/templates/index.ftl
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <title>Blog</title>
+    <link href="/css/font.css" rel="stylesheet">
+    <link href="/css/index.css" rel="stylesheet">
+</head>
+<body>
+    <h1>Index</h1>
+    <script src="/js/index.js" type="text/javascript"></script>
+</body>
+</html>
